Working sketches, details products used
I updated the comment headers to remove the original headers from Mike
McAuley, but I retained acknowledgement.
I referenced the URL for the product used from IC Station and made the
setting of the pins explicit. NOTE - it can go to a DIGITAL pin on
receive. (although it is an analog signal). Anyhoo - it worked - hello
world transmitted from one to the other.